analyzer/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
rasterio==1.3.11
geopandas==1.0.1
shapely==2.0.6
pyproj==3.7.0
numpy==2.1.3
requests==2.32.3
python-multipart==0.0.12
pystac-client==0.8.5
httpx==0.28.1
Pillow==11.0.0
scipy==1.14.1
fiona==1.10.1
python-dotenv==1.0.1
